# Fix Regression when using sum/cumsum on Groupby objects 