- [x] closes #43292
- [x] tests added / passed
- [ ] Ensure all linting tests pass, see [here](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#pre-commit) for how to run them
- [ ] whatsnew entry

NOTE: This will reduce performance significantly. Hopefully someone points me to a better check for `nan` which can be done with released GIL.