@phofl 2021-11-19
Performance was reduced through introducing Kahan. I don't think we can reduce it further. Have to find a way around the gil