@CloseChoice 2021-11-19
got around the gil. Now it's still the additional checks for `np.nan` and `-np.nan` in each iteration. But what are these failed checks about?