@mzeitlin11 2021-11-19
> got around the gil. Now it's still the additional checks for `np.nan` and `-np.nan` in each iteration. But what are these failed checks about?

Usually a build error when you see that. In this case: 
```
pandas/_libs/groupby.c:12588:18: error: ‘__pyx_v_t’ may be used uninitialized in this function [-Werror=maybe-uninitialized]
12588 |               if (__pyx_t_13) {
      |                  ^
pandas/_libs/groupby.c:12228:28: note: ‘__pyx_v_t’ was declared here
12228 |   __pyx_t_5numpy_float32_t __pyx_v_t;
      |                            ^~~~~~~~~
pandas/_libs/groupby.c: In function ‘__pyx_fuse_9__pyx_pw_6pandas_5_libs_7groupby_57group_cumsum’:
pandas/_libs/groupby.c:13372:18: error: ‘__pyx_v_t’ may be used uninitialized in this function [-Werror=maybe-uninitialized]
13372 |               if (__pyx_t_13) {
      |                  ^
pandas/_libs/groupby.c:13012:28: note: ‘__pyx_v_t’ was declared here
13012 |   __pyx_t_5numpy_float64_t __pyx_v_t;
      |                            ^~~~~~~~~
cc1: all warnings being treated as errors
```