@CloseChoice 2021-11-22
> What do the benchmarks look like here? Since we already have the regression #39622 from using kahan (and this adds decent complexity), would like to propose returning to previous simple summation for 1.3.5, which would fix both that regression and this one. Maybe we could expose kahan summation as an optional keyword instead (which a user can just know not to use if their data has infs?)

Here are the performance measures:
``` python
import numpy as np
import pandas as pd
import timeit

arr = np.ones(1000000)
s = pd.Series(np.arange(len(arr)))
s = s / 10

%timeit -r7 -n1000 s.groupby(arr).sum()
# on this branch
# 15.4 ms ± 172 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
# on 1.3.4
# 15.4 ms ± 119 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)

idx = pd.date_range(start="1/1/2000", end="1/1/2001", freq="T")
s = pd.Series(np.random.randn(len(idx)), index=idx)
%timeit -r7 -n1000 s.resample("1D").mean()
# on this branch
# 4.67 ms ± 44.4 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
# on 1.3.4
# 4.61 ms ± 35.9 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
```
So there is a performance regression of 2-3%
Maybe it would be the best option just to go back to normal summation and expose kahan's summation as a keyword. If the user decides to take the slower but numerically more stable summation, I guess he/she can affort the extra time to get the correct results for the special case with `np.inf`.