@simonjayhawkins 2021-11-27
> What do the benchmarks look like here? Since we already have the regression #39622 from using kahan (and this adds decent complexity), would like to propose returning to previous simple summation for 1.3.5, which would fix both that regression and this one. Maybe we could expose kahan summation as an optional keyword instead (which a user can just know not to use if their data has infs?)

needs a release note. adding the 1.3.5 milestone to match the linked issue.

thoughts on backporting these changes?