@CloseChoice 2021-11-28
Thanks @jbrockmendel for pointing me to the isinf function from numpy. This also influences the performance:

```python
In [1]: from asv_bench.benchmarks.groupby import *

In [2]: self = GroupByCythonAgg()

In [3]: self.setup("float64", "sum")

In [4]: %timeit -r 30 -n 100 self.time_frame_agg("float64", "sum")
# This PR
# 25.3 ms ± 377 µs per loop (mean ± std. dev. of 30 runs, 100 loops each)
# master
# 19.8 ms ± 355 µs per loop (mean ± std. dev. of 30 runs, 100 loops each)
```